# docs/swagger.yaml
basePath: /api/v1
definitions:
  ent.MyCard:
    properties:
      card_password:
        description: CardPassword holds the value of the "card_password" field.
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      taken_at:
        description: TakenAt holds the value of the "taken_at" field.
        type: string
    type: object
  ent.Notification:
    properties:
      id:
        description: ID of the ent.
        type: string
      message:
        description: Message holds the value of the "message" field.
        type: string
      read_at:
        description: ReadAt holds the value of the "read_at" field.
        type: string
      sent_at:
        description: SentAt holds the value of the "sent_at" field.
        type: string
      type:
        allOf:
        - $ref: '#/definitions/notification.Type'
        description: Type holds the value of the "type" field.
    type: object
  ent.Price:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: Description holds the value of the "description" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      name:
        description: Name holds the value of the "name" field.
        type: string
      taken_at:
        description: TakenAt holds the value of the "taken_at" field.
        type: string
    type: object
...
  ent.User:
    properties:
      birth_year:
        description: BirthYear holds the value of the "birth_year" field.
        type: integer
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      demented_among_direct_relatives:
        description: DementedAmongDirectRelatives holds the value of the "demented_among_direct_relatives"
          field.
        type: boolean
      ear_condition:
        allOf:
        - $ref: '#/definitions/user.EarCondition'
        description: EarCondition holds the value of the "ear_condition" field.
      education_level:
        allOf:
        - $ref: '#/definitions/user.EducationLevel'
        description: EducationLevel holds the value of the "education_level" field.
      eyesight_condition:
        allOf:
        - $ref: '#/definitions/user.EyesightCondition'
        description: EyesightCondition holds the value of the "eyesight_condition"
          field.
      gender:
        allOf:
        - $ref: '#/definitions/user.Gender'
        description: Gender holds the value of the "gender" field.
      head_injury_experience:
        description: HeadInjuryExperience holds the value of the "head_injury_experience"
          field.
        type: boolean
      height:
        description: Height holds the value of the "height" field.
        type: number
      id:
        description: ID of the ent.
        type: string
      marriage:
        allOf:
        - $ref: '#/definitions/user.Marriage'
        description: Marriage holds the value of the "marriage" field.
      medical_history:
        description: MedicalHistory holds the value of the "medical_history" field.
        type: string
      medication_status:
        description: MedicationStatus holds the value of the "medication_status" field.
        type: string
      occupation:
        allOf:
        - $ref: '#/definitions/user.Occupation'
        description: Occupation holds the value of the "occupation" field.
      smoking_habit:
        allOf:
        - $ref: '#/definitions/user.SmokingHabit'
        description: SmokingHabit holds the value of the "smoking_habit" field.
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      weight:
        description: Weight holds the value of the "weight" field.
        type: number
    type: object
...
  notification.Type:
    enum:
    - normal
    - mycard
    - price
    type: string
    x-enum-varnames:
    - TypeNormal
    - TypeMycard
    - TypePrice
...
  types.BaseUser:
    properties:
      birth_year:
        example: 2000
        type: integer
      demented_among_direct_relatives:
        example: false
        type: boolean
      ear_condition:
        enum:
        - normal
        - slightly_affecting_conversation
        - need_hearing_aid
        example: normal
        type: string
      education_level:
        enum:
        - elementry_school_or_below
        - middle_school
        - high_school
        - bachelor
        - master
        - doctorate
        example: doctorate
        type: string
      eyesight_condition:
        enum:
        - normal
        - slightly_affecting_reading
        - need_glasses
        example: normal
        type: string
      gender:
        example: male
        type: string
      head_injury_experience:
        example: false
        type: boolean
      height:
        example: 185.2
        type: number
      id:
        type: string
      marriage:
        enum:
        - single
        - married
        - divorced
        - widowed
        example: single
        type: string
      medical_history:
        example: none
        type: string
      medication_status:
        example: none
        type: string
      occupation:
        enum:
        - student
        - government_employee
        - service_industry
        - industry_and_commerce
        - freelancer
        - domestic
        example: student
        type: string
      smoking_habit:
        enum:
        - none
        - sometimes
        - everyday
        example: none
        type: string
      weight:
        example: 80.3
        type: number
    type: object
...
  user.EarCondition:
    enum:
    - normal
    - slightly_affecting_conversation
    - need_hearing_aid
    type: string
    x-enum-varnames:
    - EarConditionNormal
    - EarConditionSlightlyAffectingConversation
    - EarConditionNeedHearingAid
  user.EducationLevel:
    enum:
    - elementry_school_or_below
    - middle_school
    - high_school
    - bachelor
    - master
    - doctorate
    type: string
    x-enum-varnames:
    - EducationLevelElementrySchoolOrBelow
    - EducationLevelMiddleSchool
    - EducationLevelHighSchool
    - EducationLevelBachelor
    - EducationLevelMaster
    - EducationLevelDoctorate
  user.EyesightCondition:
    enum:
    - normal
    - slightly_affecting_reading
    - need_glasses
    type: string
    x-enum-varnames:
    - EyesightConditionNormal
    - EyesightConditionSlightlyAffectingReading
    - EyesightConditionNeedGlasses
  user.Gender:
    enum:
    - male
    - female
    - nonbinary
    type: string
    x-enum-varnames:
    - GenderMale
    - GenderFemale
    - GenderNonbinary
  user.Marriage:
    enum:
    - single
    - married
    - divorced
    - widowed
    type: string
    x-enum-varnames:
    - MarriageSingle
    - MarriageMarried
    - MarriageDivorced
    - MarriageWidowed
  user.Occupation:
    enum:
    - student
    - government_employee
    - service_industry
    - industry_and_commerce
    - freelancer
    - domestic
    type: string
    x-enum-varnames:
    - OccupationStudent
    - OccupationGovernmentEmployee
    - OccupationServiceIndustry
    - OccupationIndustryAndCommerce
    - OccupationFreelancer
    - OccupationDomestic
  user.SmokingHabit:
    enum:
    - none
    - sometimes
    - everyday
    type: string
    x-enum-varnames:
    - SmokingHabitNone
    - SmokingHabitSometimes
    - SmokingHabitEveryday
host: health-statistic.dechnology.com.tw
info:
  contact: {}
  description: "This is the API documentation for 「健康資料公鏈」\n\n# Usage\n\n## Flow Diagram\n![flow_diagram](https://cdn.discordapp.com/attachments/874556062815100940/1132920083174408222/App-.drawio.png)\n\n##
    API Endpoints for Frontend Application\n\nFull descriptions of all endpoints are
    described below; however, more common endpoints for a client application are described
    here. For detailed types of params, body, or response, please refer to the full
    descriptiions below. \n\n### Server Health Check - `/health_check` [**GET**]\n\nThis
    endpoint checks if the server is alive and the database is connected. It returns
    a code of 200 if ture.\n\n---\n\n### Get Registration Questionnaire - `/questionnaires/registration`
    [**GET**]\n\nFor the sake of scalibility, we make the questionnaire system a general
    one. That is, we can later create more questionnaires in our application. Upon
    initialization of our database, the registration questionnaire is created and
    assigned a fixed UUID. Before the user register, we need to have them fill in
    the registration questionnaire. This questionniare can be fetch using this endpoint.
    Note that this questionnaire correspond to question 16 to 30 since the answers
    to question 1 to 15 are part of the user data.\n\n---\n\n### User Registration
...
    some answers are part of the user data like \"birth_year\", \"gender\", and so
...
paths:
...
  /users:
    get:
      description: Get all users from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.User'
            type: array
      summary: Get Users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: '**This is a debugging endpoint and it is not meant to be consumed
        by the client side.**'
      parameters:
      - description: The user to be created
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.BaseUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
      summary: Create User
      tags:
      - User
  /users/{id}:
    delete:
      parameters:
      - description: The user's Auth0 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete User
      tags:
      - User
    get:
      description: |
        Get the user specified by the `id` path param.
      parameters:
      - description: The user's Auth0 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
      summary: Get User
      tags:
      - User
    put:
      consumes:
      - application/json
      parameters:
      - description: The user's Auth0 ID
        in: path
        name: id
        required: true
        type: string
      - description: user to be updated
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.BaseUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
      summary: Update User
      tags:
      - User
  /users/{id}/mycards:
    get:
      description: The ID field is the card number you are supposed to enter when
        redeeming the MyCards.
      parameters:
      - description: The user's Auth0 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.MyCard'
            type: array
      summary: Get All MyCards From an User
      tags:
      - User
  /users/{id}/notifications:
    get:
      description: |
        Get all history notifications of an user.
      parameters:
      - description: The user's Auth0 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/ent.Notification'
              type: array
            type: array
      summary: Get All Notifications From an User
      tags:
      - User
  /users/{id}/prices:
    get:
      parameters:
      - description: The user's Auth0 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Price'
            type: array
      summary: Get All Prices From an User
      tags:
      - User
schemes:
...
swagger: "2.0"